DA-6 Move com.smartsocietyservices to org.osgpfoundation
