+++ osgp-dto/src/main/java/org/osgpfoundation/osgp/dto/da/GetDeviceModelRequestDto.java
@@ -1,11 +1,12 @@
-package com.smartsocietyservices.osgp.dto.da;
+package org.osgpfoundation.osgp.dto.da;
 
 import java.io.Serializable;
 
-public class GetHealthStatusRequestDto implements Serializable {
+public class GetDeviceModelRequestDto implements Serializable {
     private final String deviceIdentifier;
 
-    public GetHealthStatusRequestDto( final String deviceIdentifier ) {
+    public GetDeviceModelRequestDto( final String deviceIdentifier ) {
+
         this.deviceIdentifier = deviceIdentifier;
     }
