+++ osgp-dto/src/main/java/org/osgpfoundation/osgp/dto/da/GetHealthStatusRequestDto.java
@@ -1,12 +1,11 @@
-package com.smartsocietyservices.osgp.dto.da;
+package org.osgpfoundation.osgp.dto.da;
 
 import java.io.Serializable;
 
-public class GetDeviceModelRequestDto implements Serializable {
+public class GetHealthStatusRequestDto implements Serializable {
     private final String deviceIdentifier;
 
-    public GetDeviceModelRequestDto( final String deviceIdentifier ) {
-
+    public GetHealthStatusRequestDto( final String deviceIdentifier ) {
         this.deviceIdentifier = deviceIdentifier;
     }
