+++ osgp-dto/src/main/java/org/osgpfoundation/osgp/dto/da/iec61850/PhysicalDeviceDto.java
@@ -1,4 +1,4 @@
-package com.smartsocietyservices.osgp.dto.da.iec61850;
+package org.osgpfoundation.osgp.dto.da.iec61850;
 
 import java.util.List;
