+++ osgp-ws-distributionautomation/src/main/java/org/osgpfoundation/osgp/ws/da/config/DistributionAutomationWebServiceConfig.java
@@ -5,7 +5,7 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package com.smartsocietyservices.osgp.ws.da.config;
+package org.osgpfoundation.osgp.ws.da.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
