+++ osgp-ws-distributionautomation/src/main/resources/AdHocManagement.wsdl
@@ -2,19 +2,19 @@
 <!-- Platform version: ${display.version} -->
 
 <wsdl:definitions xmlns:wsdl="http://schemas.xmlsoap.org/wsdl/"
-    xmlns:da="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/defs/2017/04"
+    xmlns:da="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/defs/2017/04"
     xmlns:common="http://www.alliander.com/schemas/osgp/common/2014/10"
     xmlns:soap="http://schemas.xmlsoap.org/wsdl/soap/"
-    xmlns:tns="http://www.smartsocietyservices.com/definitions/osgp/distributionautomation/2017/04"
-    targetNamespace="http://www.smartsocietyservices.com/definitions/osgp/distributionautomation/2017/04">
+    xmlns:tns="http://www.osgpfoundation.org/definitions/osgp/distributionautomation/2017/04"
+    targetNamespace="http://www.osgpfoundation.org/definitions/osgp/distributionautomation/2017/04">
 
     <wsdl:types>
         <xsd:schema
-            targetNamespace="http://www.smartsocietyservices.com/definitions/osgp/distributionautomation/imports"
+            targetNamespace="http://www.osgpfoundation.org/definitions/osgp/distributionautomation/imports"
             xmlns:xsd="http://www.w3.org/2001/XMLSchema">
             
             <xsd:import
-                namespace="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/defs/2017/04"
+                namespace="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/defs/2017/04"
                 schemaLocation="schemas/distributionautomation.xsd" />
         </xsd:schema>
     </wsdl:types>
