+++ osgp-ws-distributionautomation/src/main/resources/schemas/common.xsd
@@ -1,4 +1,4 @@
-<!-- Copyright 2015 Smart Society Services B.V. Licensed under the Apache 
+<!-- Copyright 2017 Smart Society Services B.V. Licensed under the Apache
   License, Version 2.0 (the "License"); you may not use this file except in 
   compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 -->
 <xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema"
