+++ osgp-ws-distributionautomation/src/main/resources/schemas/distributionautomation.xjb
@@ -6,14 +6,14 @@
   <jxb:bindings schemaLocation="common.xsd">
     <jxb:schemaBindings>
       <jxb:package
-        name="com.smartsocietyservices.osgp.adapter.ws.schema.distributionautomation.common"></jxb:package>
+        name="org.osgpfoundation.osgp.adapter.ws.schema.distributionautomation.common"></jxb:package>
     </jxb:schemaBindings>
   </jxb:bindings>
 
   <jxb:bindings schemaLocation="distributionautomation.xsd">
     <jxb:schemaBindings>
       <jxb:package
-        name="com.smartsocietyservices.osgp.adapter.ws.schema.distributionautomation.generic"></jxb:package>
+        name="org.osgpfoundation.osgp.adapter.ws.schema.distributionautomation.generic"></jxb:package>
     </jxb:schemaBindings>
   </jxb:bindings>
