+++ osgp-ws-distributionautomation/src/main/resources/schemas/distributionautomation.xsd
@@ -1,5 +1,5 @@
 <?xml version="1.1" encoding="UTF-8"?>
-<!-- Copyright 2015 Smart Society Services B.V. 
+<!-- Copyright 2017 Smart Society Services B.V.
     
     Licensed under the Apache License, Version 2.0 (the "License"); 
     you may not use this file except in compliance with the License. 
@@ -8,9 +8,9 @@
 -->
 <xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema"
     xmlns:common="http://www.alliander.com/schemas/osgp/common/2014/10"
-    xmlns:da="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/defs/2017/04"
+    xmlns:da="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/defs/2017/04"
     attributeFormDefault="unqualified" elementFormDefault="qualified"
-    targetNamespace="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/defs/2017/04">
+    targetNamespace="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/defs/2017/04">
 
     <xsd:import namespace="http://www.alliander.com/schemas/osgp/common/2014/10"
         schemaLocation="common.xsd"/>
