+++ osgp-ws-distributionautomation/src/main/resources/schemas/notification.xsd
@@ -1,11 +1,11 @@
-<!-- Copyright 2015 Smart Society Services B.V. Licensed under the Apache 
+<!-- Copyright 2017 Smart Society Services B.V. Licensed under the Apache
   License, Version 2.0 (the "License"); you may not use this file except in 
   compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 -->
 <xsd:schema xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-  xmlns:tns="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/notification/2017/04"
+  xmlns:tns="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/notification/2017/04"
   xmlns:common="http://www.alliander.com/schemas/osgp/common/2014/10"
   attributeFormDefault="unqualified" elementFormDefault="qualified"
-  targetNamespace="http://www.smartsocietyservices.com/schemas/osgp/distributionautomation/notification/2017/04">
+  targetNamespace="http://www.osgpfoundation.org/schemas/osgp/distributionautomation/notification/2017/04">
 
   <xsd:import namespace="http://www.alliander.com/schemas/osgp/common/2014/10"
     schemaLocation="common.xsd"/>
